move use case to common entities
