+++ src/main/java/com/truelayer/java/paymentsproviders/entities/VrpCommercialCapabilities.java
@@ -2,7 +2,7 @@ package com.truelayer.java.paymentsproviders.entities;
 
 import com.truelayer.java.entities.ProviderAvailability;
 import com.truelayer.java.payments.entities.ReleaseChannel;
-import com.truelayer.java.payments.entities.UseCase;
+import com.truelayer.java.entities.UseCase;
 import lombok.Value;
 
 import java.util.List;
